Add `Sia::Decorator::ColorManipulations`
